+++ packages/ui/button/stories/icloud.stories.tsx
@@ -5,6 +5,7 @@ import Button from '../src';
 
 export default {
     title: 'Button',
+    component: Button,
 };
 
 export const Demo = () => {