+++ packages/ui/button/stories/osui.stories.tsx
@@ -4,6 +4,7 @@ import Button from '../src';
 
 export default {
     title: 'Button',
+    component: Button,
 };
 
 export const Demo = () => {